version 0.1.1 (some bug fixes)
If a new stanmodel is compiled, a message is now printed which indicates
this (still working on how to be able to use the new stanmodel a second
time without closing).  Fixed bug with NA interactions which involved
two unordered factors both with NA values and one or more additional
unordered factors.