changed default for estimate_scale_beta
In bmer_control, the default for estimate_scale_beta was previously "P",
which uses the "no" behavior when P < 10 and the "yes" behavior when P
>= 10.  Putting a hyperprior on the fixed effect t-dist scale can create
a prohibitively tight prior when there are many interactions involved.
The functionality of the function hasn't changed, but the recommended
default is to use a fixed scale parameter.